requirements.txt
flask
flask-socketio
Flask-Session
pillow
wordcloud